+++ toolium/driver_wrapper.py
@@ -208,4 +208,4 @@ class DriverWrapper(object):
         :returns: true if the browser is maximizable
         """
         browser_name = self.config.get('Browser', 'browser').split('-')[0]
-        return not self.is_mobile_test() and browser_name != 'opera' and browser_name != 'edge'
+        return not self.is_mobile_test() and browser_name != 'edge'